A simple roguelike in Ada to run on the Windows Command Prompt or on Powershell.
The first commit has a main procedure that just uses the win32 API to
measure the size of the console screen in characters. This will later be
moved to its own function and the screen size used to generate the rest
of the roguelike text graphics screen.

Right now src has a flat structure, but in the future helper files will
be sorted by directory.

.gitignore has just enough to prevent this first commit from adding
unnecessary files, but will have several, more general excludes in the
future.

In future commits, dedicated documentation should be added. For now the
documentation is just comments.