Moved the code that measure the size of the console from the main body to a separate console package.
Created win and console directories to organize the new console code and
the win32 ffi code. For now, the only file in src root folder will be
main.adb.

Since the win32 ffi code is used in the Console code, but isn't part of
the Console directory tree, this may cause some issues in the future. At
least right now, autocomplete doesn't work with the current structure. I
may want to find a way to reorganize this in the future.